Add A/B test for "Reply-To"
This maybe isn't so compelling as a feature.  But we kinda get boxed-in,
because the "From" field gets "A/B", and there's conditional sync'ing
between "From"<=>"Reply-To" (based on the `replyTo` setting).

This makes it safer to handle reconfigurations of `replyTo`, as in:

1. Start with replyTo=false (default)
2. Make a mailing. Split the "From". Save draft.
3. Reconfigure replyTo=true
4. Resume draft or re-use mailing with A/B test on "From"

Without this patch, the UI would present a singular "Reply-To", but the
underlying "variants.X.replyto_email" would have two (invisible/uneditable)
overrides.

With this patch, in that scenario, the values of "variants.X.replyto_email"
would be apparent.